Split IAP extensions
