+++ Passepartout/Library/Sources/CommonLibrary/IAP/AppFeatureRequiring.swift
@@ -24,90 +24,7 @@
 //
 
 import Foundation
-import PassepartoutKit
 
 public protocol AppFeatureRequiring {
     var features: Set<AppFeature> { get }
-}
-
-// MARK: - Profile
-
-extension Profile: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        let builders = activeModules.compactMap { module in
-            guard let builder = module.moduleBuilder() else {
-                fatalError("Cannot produce ModuleBuilder from Module: \(module)")
-            }
-            return builder
-        }
-        return builders.features
-    }
-}
-
-extension Array: AppFeatureRequiring where Element == any ModuleBuilder {
-    public var features: Set<AppFeature> {
-        let requirements = compactMap { builder in
-            guard let requiring = builder as? AppFeatureRequiring else {
-                fatalError("ModuleBuilder does not implement AppFeatureRequiring: \(builder)")
-            }
-            return requiring
-        }
-        return Set(requirements.flatMap(\.features))
-    }
-}
-
-// MARK: - Modules
-
-extension DNSModule.Builder: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        [.dns]
-    }
-}
-
-extension HTTPProxyModule.Builder: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        [.httpProxy]
-    }
-}
-
-extension IPModule.Builder: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        [.routing]
-    }
-}
-
-extension OnDemandModule.Builder: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        guard isEnabled else {
-            return []
-        }
-        return policy != .any ? [.onDemand] : []
-    }
-}
-
-extension OpenVPNModule.Builder: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        var list: Set<AppFeature> = []
-        providerId?.features.forEach {
-            list.insert($0)
-        }
-        if isInteractive {
-            list.insert(.interactiveLogin)
-        }
-        return list
-    }
-}
-
-extension WireGuardModule.Builder: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        providerId?.features ?? []
-    }
-}
-
-// MARK: - Providers
-
-extension ProviderID: AppFeatureRequiring {
-    public var features: Set<AppFeature> {
-        self != .oeck ? [.providers] : []
-    }
 }
